requirements.txt
pip-tools
matplotlib
ultralytics
psycopg2-binary
pycryptodome
django-admin-rangefilter
pyqtgraph
numpy
gdown
Pillow
opencv-python
psutil
PySide6
attrs>=17.2.0
qrcode
urllib3<2.1,>=1.25.4
certifi
idna
onvif-zeep
scapy
requests
tensorboard
pyinstaller
cx_Freeze
auto_py_to_exe
cython
nuitka